Fix broken paths after moving files and install Sass
